+++ src/index.ts
@@ -1,6 +1,9 @@
 export { AccountLinksClient } from "./api/accountLinks";
 export { AccountsClient } from "./api/accounts";
 export { ApplePayClient } from "./api/applePay";
+export { BalancesClient } from "./api/balances";
+export { BalanceTransactionsClient } from "./api/balanceTransactions";
+export { ConversionsClient } from "./api/conversions";
 export { CustomersClient } from "./api/customers";
 export { DisputesClient } from "./api/disputes";
 export { EventsClient } from "./api/events";
@@ -20,5 +23,10 @@ export { SubscriptionsClient } from "./api/subscriptions";
 export { TransfersClient } from "./api/transfers";
 export { WebhooksClient } from "./api/webhooks";
 export { Config } from "./types/config";
+export { ErrorElement, RyftError, RyftErrorResponse } from "./types/errors";
 export { determineBaseUrl } from "./utils/url";
-export { Ryft } from "./client"; 
+export { WebhookSignatureVerifier } from "./utils/webhookSignatureVerifier";
+export { Ryft } from "./client";
+
+export * from "./types/conversions/conversions";
+export * from "./types/conversions/req";