+++ test/api/conversions/mockData/mockCreateConversionReq.ts
@@ -1,0 +1,25 @@
+import { CreateConversionRequest } from '../../../../src';
+
+export const mockCreateConversionReq: CreateConversionRequest = {
+  sell: {
+    currency: 'GBP',
+    amount: 500,
+  },
+  buy: {
+    currency: 'EUR',
+  },
+  termAgreement: true,
+  reason: 'Paying EUR suppliers',
+};
+
+export const mockCreateConversionReq_FixedSell: CreateConversionRequest = {
+  sell: {
+    currency: 'GBP',
+    amount: 500,
+  },
+  buy: {
+    currency: 'EUR',
+  },
+  termAgreement: true,
+  fixedSide: 'Sell',
+};